auth_service/requirement.txt
fastapi~=0.111.0
uvicorn~=0.30.1
python-dotenv~=1.0.1
motor~=3.4.0
passlib~=1.7.4
bcrypt~=4.1.3
pyjwt~=2.8.0
pydantic-settings~=2.3.3
fastapi-mail~=1.4.1
pillow~=10.3.0
python-multipart~=0.0.9

